fix: Various Fixes
-Execute threshold now properly casts to floats
-End turn fixes
-Card changes

+++ Assets/Scripts/Combat/Characters/CharacterHealth.cs
@@ -99,7 +99,7 @@ namespace Necropanda
             return trueValue;
         }
 
-        public float GetHealthPercentage() { return health / maxHealth; }
+        public float GetHealthPercentage() { return (float)health / (float)maxHealth; }
 
         void UpdateHealthUI()
         {
